comms_only is working. real basic.
