+++ platformio.ini
@@ -10,8 +10,10 @@
 
 [platformio]
 ; default_envs = esp32dev.debug
-default_envs = esp32dev.release
+; default_envs = esp32dev.release
+default_envs = esp32dev.comms_only
 
+;------------------------------------------------
 [env]
 
 platform = espressif32
@@ -31,12 +33,10 @@ lib_deps =
   https://github.com/nRF24/RF24Network.git
   https://github.com/skelstar/NRF24L01.Library.git
 
-monitor_speed = 115200
-monitor_port = COM12
-upload_port = COM12
-
 ;-----------------------------------------------
 [env:esp32dev.debug]
+
+src_filter = ${dir.main}
 
 build_flags = 
   -D USE_TEST_VALUES
@@ -51,6 +51,8 @@ build_flags =
 ;-----------------------------------------------
 [env:esp32dev.release]
 
+src_filter = ${dir.main}
+
 build_flags = 
   -D USE_TEST_VALUES
   ; -D DEBUG_PRINT_FSM_EVENT
@@ -59,3 +61,36 @@ build_flags =
   ; -D DEBUG_PRINT_SENT_TO_CONTROLLER
   ; -D USING_BUTTONS
   -D LOG_RETRIES
+;-------------------------------------------------------
+
+[env:esp32dev.comms_only]
+
+src_filter = ${dir.comms_only}
+
+build_flags = 
+  ; -D USE_TEST_VALUES
+  ; -D DEBUG_PRINT_FSM_EVENT
+  ; -D DEBUG_PRINT_THROTTLE_VALUE
+  ; -D PRINT_STATE_NAME
+  ; -D DEBUG_PRINT_SENT_TO_CONTROLLER
+  ; -D USING_BUTTONS
+  -D LOG_RETRIES
+
+lib_deps = 
+  https://github.com/lewisxhe/Button2.git
+  https://github.com/pfeerick/elapsedMillis.git
+  https://github.com/MattFryer/Smoothed.git
+  
+  https://github.com/skelstar/arduino-fsm.git 
+  https://github.com/skelstar/VESC_Comms.git
+  https://github.com/skelstar/Esk8.Board.VescComms.git#master
+  https://github.com/skelstar/Arduino-Debugging.git
+
+  https://github.com/nRF24/RF24Network.git
+  https://github.com/skelstar/NRF24L01.Library.git#master
+
+
+;-------------------------------------------------------
+[dir]
+main = +<main/> -<comms_only/>
+comms_only = +<comms_only/>
